Split payload and state into their own files
