+++ tp/src/handler.rs
@@ -11,9 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-use std;
 
-use protobuf;
 use protobuf::RepeatedField;
 
 use crypto::digest::Digest;
@@ -22,33 +20,26 @@ use crypto::sha2::Sha512;
 use wasm_executor::wasm_module::WasmModule;
 
 use sawtooth_sdk::messages::processor::TpProcessRequest;
-use sawtooth_sdk::messages::setting::Setting;
 use sawtooth_sdk::processor::handler::ApplyError;
 use sawtooth_sdk::processor::handler::TransactionContext;
 use sawtooth_sdk::processor::handler::TransactionHandler;
 
-use addressing::{
-    compute_agent_address, compute_org_address, compute_smart_permission_address,
-    get_sawtooth_admins_address, make_contract_address, make_contract_registry_address,
-    make_namespace_registry_address,
-};
-
-use protos::contract::{Contract, ContractList};
-use protos::contract_registry::{ContractRegistry, ContractRegistryList, ContractRegistry_Version};
+use protos::contract::Contract;
+use protos::contract_registry::{ContractRegistry, ContractRegistry_Version};
 use protos::namespace_registry::{
-    NamespaceRegistry, NamespaceRegistryList, NamespaceRegistry_Permission,
+    NamespaceRegistry, NamespaceRegistry_Permission,
 };
 use protos::payload::{
     CreateContractAction, CreateContractRegistryAction, CreateNamespaceRegistryAction,
     CreateNamespaceRegistryPermissionAction, CreateSmartPermissionAction, DeleteContractAction,
     DeleteContractRegistryAction, DeleteNamespaceRegistryAction,
     DeleteNamespaceRegistryPermissionAction, DeleteSmartPermissionAction, ExecuteContractAction,
-    SabrePayload, SabrePayload_Action, UpdateContractRegistryOwnersAction,
-    UpdateNamespaceRegistryOwnersAction, UpdateSmartPermissionAction,
+    UpdateContractRegistryOwnersAction, UpdateNamespaceRegistryOwnersAction,
+    UpdateSmartPermissionAction,
 };
-use protos::smart_permission::{
-    Agent, AgentList, Organization, OrganizationList, SmartPermission, SmartPermissionList,
-};
+use protos::smart_permission::SmartPermission;
+use state::SabreState;
+use payload::{Action, SabreRequestPayload};
 
 /// The namespace registry prefix for global state (00ec00)
 const NAMESPACE_REGISTRY_PREFIX: &'static str = "00ec00";
@@ -58,873 +49,6 @@ const CONTRACT_REGISTRY_PREFIX: &'static str = "00ec01";
 
 /// The contract prefix for global state (00ec02)
 const CONTRACT_PREFIX: &'static str = "00ec02";
-
-#[derive(Debug, Clone)]
-enum Action {
-    CreateContract(CreateContractAction),
-    DeleteContract(DeleteContractAction),
-    ExecuteContract(ExecuteContractAction),
-    CreateContractRegistry(CreateContractRegistryAction),
-    DeleteContractRegistry(DeleteContractRegistryAction),
-    UpdateContractRegistryOwners(UpdateContractRegistryOwnersAction),
-    CreateNamespaceRegistry(CreateNamespaceRegistryAction),
-    DeleteNamespaceRegistry(DeleteNamespaceRegistryAction),
-    UpdateNamespaceRegistryOwners(UpdateNamespaceRegistryOwnersAction),
-    CreateNamespaceRegistryPermission(CreateNamespaceRegistryPermissionAction),
-    DeleteNamespaceRegistryPermission(DeleteNamespaceRegistryPermissionAction),
-    CreateSmartPermission(CreateSmartPermissionAction),
-    UpdateSmartPermission(UpdateSmartPermissionAction),
-    DeleteSmartPermission(DeleteSmartPermissionAction),
-}
-
-impl std::fmt::Display for Action {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match *self {
-            Action::CreateContract(_) => write!(f, "Action: Create Contract"),
-            Action::DeleteContract(_) => write!(f, "Action: Delete Contract"),
-            Action::ExecuteContract(_) => write!(f, "Action: Execute Contract"),
-            Action::CreateContractRegistry(_) => write!(f, "Action: Create Contract Registry"),
-            Action::DeleteContractRegistry(_) => write!(f, "Action: Delete Contract Registry"),
-            Action::UpdateContractRegistryOwners(_) => {
-                write!(f, "Action: Update Contract Registry Owners")
-            }
-            Action::CreateNamespaceRegistry(_) => write!(f, "Action: Create Namespace Registry"),
-            Action::DeleteNamespaceRegistry(_) => write!(f, "Action: Delete Namespace Registry"),
-            Action::UpdateNamespaceRegistryOwners(_) => {
-                write!(f, "Action: Update Namespace Registry Owners")
-            }
-            Action::CreateNamespaceRegistryPermission(_) => {
-                write!(f, "Create Namespace Registry Permission")
-            }
-            Action::DeleteNamespaceRegistryPermission(_) => {
-                write!(f, "Delete Namespace Registry Permission")
-            }
-            Action::CreateSmartPermission(_) => write!(f, "Create smart permission"),
-            Action::UpdateSmartPermission(_) => write!(f, "Update smart permission"),
-            Action::DeleteSmartPermission(_) => write!(f, "Delete smart permission"),
-        }
-    }
-}
-
-struct SabreRequestPayload {
-    action: Action,
-}
-
-impl SabreRequestPayload {
-    pub fn new(payload: &[u8]) -> Result<Option<SabreRequestPayload>, ApplyError> {
-        let payload: SabrePayload = match protobuf::parse_from_bytes(payload) {
-            Ok(payload) => payload,
-            Err(_) => {
-                return Err(ApplyError::InvalidTransaction(String::from(
-                    "Cannot deserialize payload",
-                )));
-            }
-        };
-
-        let sabre_action = payload.get_action();
-        let action = match sabre_action {
-            SabrePayload_Action::CREATE_CONTRACT => {
-                let create_contract = payload.get_create_contract();
-                if create_contract.get_name().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract name cannot be an empty string",
-                    )));
-                }
-                if create_contract.get_version().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract version cannot be an empty string",
-                    )));
-                }
-                if create_contract.get_inputs().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract inputs cannot be an empty",
-                    )));
-                }
-                if create_contract.get_outputs().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract outputs cannot be an empty",
-                    )));
-                }
-                if create_contract.get_contract().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract bytes cannot be an empty",
-                    )));
-                }
-                Action::CreateContract(create_contract.clone())
-            }
-            SabrePayload_Action::DELETE_CONTRACT => {
-                let delete_contract = payload.get_delete_contract();
-                if delete_contract.get_name().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract name cannot be an empty string",
-                    )));
-                }
-                if delete_contract.get_version().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract version cannot be an empty string",
-                    )));
-                }
-                Action::DeleteContract(delete_contract.clone())
-            }
-            SabrePayload_Action::EXECUTE_CONTRACT => {
-                let execute_contract = payload.get_execute_contract();
-                if execute_contract.get_name().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract name cannot be an empty string",
-                    )));
-                }
-                if execute_contract.get_version().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract version cannot be an empty string",
-                    )));
-                }
-                if execute_contract.get_inputs().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract inputs cannot be an empty",
-                    )));
-                }
-                if execute_contract.get_outputs().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract outputs cannot be an empty",
-                    )));
-                }
-                if execute_contract.get_payload().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract payload cannot be an empty",
-                    )));
-                }
-                Action::ExecuteContract(execute_contract.clone())
-            }
-            SabrePayload_Action::CREATE_CONTRACT_REGISTRY => {
-                let create_contract_registry = payload.get_create_contract_registry();
-                if create_contract_registry.get_name().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract Registry name cannot be an empty string",
-                    )));
-                }
-                if create_contract_registry.get_owners().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract Registry owners cannot be an empty",
-                    )));
-                }
-                Action::CreateContractRegistry(create_contract_registry.clone())
-            }
-            SabrePayload_Action::DELETE_CONTRACT_REGISTRY => {
-                let delete_contract_registry = payload.get_delete_contract_registry();
-                if delete_contract_registry.get_name().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract Registry name cannot be an empty string",
-                    )));
-                };
-                Action::DeleteContractRegistry(delete_contract_registry.clone())
-            }
-            SabrePayload_Action::UPDATE_CONTRACT_REGISTRY_OWNERS => {
-                let update_contract_registry_owners = payload.get_update_contract_registry_owners();
-                if update_contract_registry_owners.get_name().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract Registry name cannot be an empty string",
-                    )));
-                }
-                if update_contract_registry_owners.get_owners().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract Registry owners cannot be an empty",
-                    )));
-                }
-                Action::UpdateContractRegistryOwners(update_contract_registry_owners.clone())
-            }
-            SabrePayload_Action::CREATE_NAMESPACE_REGISTRY => {
-                let create_namespace_registry = payload.get_create_namespace_registry();
-                if create_namespace_registry.get_namespace().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Namespace Registry namespace cannot be an empty string",
-                    )));
-                }
-                if create_namespace_registry.get_owners().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Namespace owners cannot be an empty",
-                    )));
-                }
-                Action::CreateNamespaceRegistry(create_namespace_registry.clone())
-            }
-            SabrePayload_Action::DELETE_NAMESPACE_REGISTRY => {
-                let delete_namespace_registry = payload.get_delete_namespace_registry();
-                if delete_namespace_registry.get_namespace().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Namespace Registry namespace cannot be an empty string",
-                    )));
-                }
-                Action::DeleteNamespaceRegistry(delete_namespace_registry.clone())
-            }
-            SabrePayload_Action::UPDATE_NAMESPACE_REGISTRY_OWNERS => {
-                let update_namespace_registry_owners =
-                    payload.get_update_namespace_registry_owners();
-                if update_namespace_registry_owners.get_namespace().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Namespace Registry namespace cannot be an empty string",
-                    )));
-                }
-                if update_namespace_registry_owners.get_owners().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Namespace owners cannot be an empty",
-                    )));
-                }
-                Action::UpdateNamespaceRegistryOwners(update_namespace_registry_owners.clone())
-            }
-            SabrePayload_Action::CREATE_NAMESPACE_REGISTRY_PERMISSION => {
-                let create_namespace_registry_permission =
-                    payload.get_create_namespace_registry_permission();
-                if create_namespace_registry_permission
-                    .get_namespace()
-                    .is_empty()
-                {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Namespace Registry namespace cannot be an empty string",
-                    )));
-                }
-                if create_namespace_registry_permission
-                    .get_contract_name()
-                    .is_empty()
-                {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract name cannot be an empty string",
-                    )));
-                }
-                Action::CreateNamespaceRegistryPermission(
-                    create_namespace_registry_permission.clone(),
-                )
-            }
-            SabrePayload_Action::DELETE_NAMESPACE_REGISTRY_PERMISSION => {
-                let delete_namespace_registry_permission =
-                    payload.get_delete_namespace_registry_permission();
-                if delete_namespace_registry_permission
-                    .get_namespace()
-                    .is_empty()
-                {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Namespace Registry namespace cannot be an empty string",
-                    )));
-                }
-                if delete_namespace_registry_permission
-                    .get_contract_name()
-                    .is_empty()
-                {
-                    return Err(ApplyError::InvalidTransaction(String::from(
-                        "Contract name cannot be an empty string",
-                    )));
-                }
-                Action::DeleteNamespaceRegistryPermission(
-                    delete_namespace_registry_permission.clone(),
-                )
-            }
-            SabrePayload_Action::CREATE_SMART_PERMISSION => {
-                let action = payload.get_create_smart_permission();
-
-                if action.get_org_id().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(
-                        "Organization ID required".into(),
-                    ));
-                }
-
-                if action.get_name().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(
-                        "Smart permission name required".into(),
-                    ));
-                }
-
-                if action.get_function().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(
-                        "Function body required".into(),
-                    ));
-                }
-
-                Action::CreateSmartPermission(action.clone())
-            }
-            SabrePayload_Action::UPDATE_SMART_PERMISSION => {
-                let action = payload.get_update_smart_permission();
-
-                if action.get_org_id().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(
-                        "Organization ID required".into(),
-                    ));
-                }
-
-                if action.get_name().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(
-                        "Smart permission name required".into(),
-                    ));
-                }
-
-                if action.get_function().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(
-                        "Function body required".into(),
-                    ));
-                }
-
-                Action::UpdateSmartPermission(action.clone())
-            }
-            SabrePayload_Action::DELETE_SMART_PERMISSION => {
-                let action = payload.get_delete_smart_permission();
-
-                if action.get_org_id().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(
-                        "Organization ID required".into(),
-                    ));
-                }
-
-                if action.get_name().is_empty() {
-                    return Err(ApplyError::InvalidTransaction(
-                        "Smart permission name required".into(),
-                    ));
-                }
-
-                Action::DeleteSmartPermission(action.clone())
-            }
-            SabrePayload_Action::ACTION_UNSET => {
-                return Err(ApplyError::InvalidTransaction(String::from(
-                    "Action is not set",
-                )));
-            }
-        };
-        Ok(Some(SabreRequestPayload { action: action }))
-    }
-
-    pub fn get_action(&self) -> Action {
-        self.action.clone()
-    }
-}
-
-pub struct SabreState<'a> {
-    context: &'a mut dyn TransactionContext,
-}
-
-impl<'a> SabreState<'a> {
-    pub fn new(context: &'a mut dyn TransactionContext) -> SabreState {
-        SabreState { context: context }
-    }
-
-    pub fn context(&mut self) -> &mut dyn TransactionContext {
-        self.context
-    }
-
-    pub fn get_admin_setting(&mut self) -> Result<Option<Setting>, ApplyError> {
-        let address = get_sawtooth_admins_address()?;
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let setting: Setting = match protobuf::parse_from_bytes(packed.as_slice()) {
-                    Ok(setting) => setting,
-                    Err(err) => {
-                        return Err(ApplyError::InternalError(format!(
-                            "Cannot deserialize setting: {:?}",
-                            err,
-                        )));
-                    }
-                };
-                Ok(Some(setting))
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn get_contract(
-        &mut self,
-        name: &str,
-        version: &str,
-    ) -> Result<Option<Contract>, ApplyError> {
-        let address = make_contract_address(name, version)?;
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let contracts: ContractList = match protobuf::parse_from_bytes(packed.as_slice()) {
-                    Ok(contracts) => contracts,
-                    Err(err) => {
-                        return Err(ApplyError::InternalError(format!(
-                            "Cannot deserialize contract list: {:?}",
-                            err,
-                        )));
-                    }
-                };
-
-                for contract in contracts.get_contracts() {
-                    if contract.name == name {
-                        return Ok(Some(contract.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_contract(
-        &mut self,
-        name: &str,
-        version: &str,
-        new_contract: Contract,
-    ) -> Result<(), ApplyError> {
-        let address = make_contract_address(name, version)?;
-        let d = self.context.get_state_entry(&address)?;
-        let mut contract_list = match d {
-            Some(packed) => match protobuf::parse_from_bytes(packed.as_slice()) {
-                Ok(contracts) => contracts,
-                Err(err) => {
-                    return Err(ApplyError::InternalError(format!(
-                        "Cannot deserialize contract list: {}",
-                        err,
-                    )));
-                }
-            },
-            None => ContractList::new(),
-        };
-        // remove old contract if it exists and sort the contracts by name
-        let contracts = contract_list.get_contracts().to_vec();
-        let mut index = None;
-        let mut count = 0;
-        for contract in contracts.clone() {
-            if contract.name == name {
-                index = Some(count);
-                break;
-            }
-            count = count + 1;
-        }
-
-        match index {
-            Some(x) => {
-                contract_list.contracts.remove(x);
-            }
-            None => (),
-        };
-        contract_list.contracts.push(new_contract);
-        contract_list.contracts.sort_by_key(|c| c.clone().name);
-        let serialized = match protobuf::Message::write_to_bytes(&contract_list) {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize contract list",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-
-    pub fn delete_contract(&mut self, name: &str, version: &str) -> Result<(), ApplyError> {
-        let address = make_contract_address(name, version)?;
-        let d = self.context.delete_state_entry(&address)?;
-        let deleted = match d {
-            Some(deleted) => deleted,
-            None => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot delete contract",
-                )));
-            }
-        };
-        if deleted != address {
-            return Err(ApplyError::InternalError(String::from(
-                "Cannot delete contract",
-            )));
-        };
-        Ok(())
-    }
-
-    pub fn get_contract_registry(
-        &mut self,
-        name: &str,
-    ) -> Result<Option<ContractRegistry>, ApplyError> {
-        let address = make_contract_registry_address(name)?;
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let contract_registries: ContractRegistryList =
-                    match protobuf::parse_from_bytes(packed.as_slice()) {
-                        Ok(contract_registries) => contract_registries,
-                        Err(err) => {
-                            return Err(ApplyError::InternalError(format!(
-                                "Cannot deserialize contract registry list: {:?}",
-                                err,
-                            )));
-                        }
-                    };
-
-                for contract_registry in contract_registries.get_registries() {
-                    if contract_registry.name == name {
-                        return Ok(Some(contract_registry.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_contract_registry(
-        &mut self,
-        name: &str,
-        new_contract_registry: ContractRegistry,
-    ) -> Result<(), ApplyError> {
-        let address = make_contract_registry_address(name)?;
-        let d = self.context.get_state_entry(&address)?;
-        let mut contract_registry_list = match d {
-            Some(packed) => match protobuf::parse_from_bytes(packed.as_slice()) {
-                Ok(contract_registries) => contract_registries,
-                Err(err) => {
-                    return Err(ApplyError::InternalError(format!(
-                        "Cannot deserialize contract registry list: {}",
-                        err,
-                    )));
-                }
-            },
-            None => ContractRegistryList::new(),
-        };
-        // remove old agent if it exists and sort the contract regisitries by name
-        let contract_registries = contract_registry_list.get_registries().to_vec();
-        let mut index = None;
-        let mut count = 0;
-        for contract_registry in contract_registries.clone() {
-            if contract_registry.name == name {
-                index = Some(count);
-                break;
-            }
-            count = count + 1;
-        }
-
-        match index {
-            Some(x) => {
-                contract_registry_list.registries.remove(x);
-            }
-            None => (),
-        };
-        contract_registry_list
-            .registries
-            .push(new_contract_registry);
-        contract_registry_list
-            .registries
-            .sort_by_key(|c| c.clone().name);
-        let serialized = match protobuf::Message::write_to_bytes(&contract_registry_list) {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize contract registry list",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-
-    pub fn delete_contract_registry(&mut self, name: &str) -> Result<(), ApplyError> {
-        let address = make_contract_registry_address(name)?;
-        let d = self.context.delete_state_entry(&address)?;
-        let deleted = match d {
-            Some(deleted) => deleted,
-            None => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot delete contract registry",
-                )));
-            }
-        };
-        if deleted != address {
-            return Err(ApplyError::InternalError(String::from(
-                "Cannot delete contract registry",
-            )));
-        };
-        Ok(())
-    }
-
-    pub fn get_namespace_registry(
-        &mut self,
-        namespace: &str,
-    ) -> Result<Option<NamespaceRegistry>, ApplyError> {
-        let address = make_namespace_registry_address(namespace)?;
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let namespace_registries: NamespaceRegistryList =
-                    match protobuf::parse_from_bytes(packed.as_slice()) {
-                        Ok(namespace_registries) => namespace_registries,
-                        Err(err) => {
-                            return Err(ApplyError::InternalError(format!(
-                                "Cannot deserialize namespace registry list: {:?}",
-                                err,
-                            )));
-                        }
-                    };
-
-                for namespace_registry in namespace_registries.get_registries() {
-                    if namespace_registry.namespace == namespace {
-                        return Ok(Some(namespace_registry.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn get_namespace_registries(
-        &mut self,
-        namespace: &str,
-    ) -> Result<Option<NamespaceRegistryList>, ApplyError> {
-        let address = make_namespace_registry_address(namespace)?;
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let namespace_registries: NamespaceRegistryList =
-                    match protobuf::parse_from_bytes(packed.as_slice()) {
-                        Ok(namespace_registries) => namespace_registries,
-                        Err(err) => {
-                            return Err(ApplyError::InternalError(format!(
-                                "Cannot deserialize namespace registry list: {:?}",
-                                err,
-                            )));
-                        }
-                    };
-                Ok(Some(namespace_registries))
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_namespace_registry(
-        &mut self,
-        namespace: &str,
-        new_namespace_registry: NamespaceRegistry,
-    ) -> Result<(), ApplyError> {
-        let address = make_namespace_registry_address(namespace)?;
-        let d = self.context.get_state_entry(&address)?;
-        let mut namespace_registry_list = match d {
-            Some(packed) => match protobuf::parse_from_bytes(packed.as_slice()) {
-                Ok(namespace_registries) => namespace_registries,
-                Err(err) => {
-                    return Err(ApplyError::InternalError(format!(
-                        "Cannot deserialize namespace registry list: {}",
-                        err,
-                    )));
-                }
-            },
-            None => NamespaceRegistryList::new(),
-        };
-        // remove old agent if it exists and sort the namespace regisitries by namespace
-        let namespace_registries = namespace_registry_list.get_registries().to_vec();
-        let mut index = None;
-        let mut count = 0;
-        for namespace_registry in namespace_registries.clone() {
-            if namespace_registry.namespace == namespace {
-                index = Some(count);
-                break;
-            }
-            count = count + 1;
-        }
-
-        match index {
-            Some(x) => {
-                namespace_registry_list.registries.remove(x);
-            }
-            None => (),
-        };
-        namespace_registry_list
-            .registries
-            .push(new_namespace_registry);
-        namespace_registry_list
-            .registries
-            .sort_by_key(|nr| nr.clone().namespace);
-        let serialized = match protobuf::Message::write_to_bytes(&namespace_registry_list) {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize namespace registry list",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-
-    pub fn delete_namespace_registry(&mut self, namespace: &str) -> Result<(), ApplyError> {
-        let address = make_namespace_registry_address(namespace)?;
-        let d = self.context.delete_state_entry(&address)?;
-        let deleted = match d {
-            Some(deleted) => deleted,
-            None => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot delete namespace registry",
-                )));
-            }
-        };
-        if deleted != address {
-            return Err(ApplyError::InternalError(String::from(
-                "Cannot delete namespace registry",
-            )));
-        };
-        Ok(())
-    }
-
-    pub fn get_smart_permission(
-        &mut self,
-        org_id: &str,
-        name: &str,
-    ) -> Result<Option<SmartPermission>, ApplyError> {
-        let address = compute_smart_permission_address(org_id, name);
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let smart_permissions: SmartPermissionList =
-                    match protobuf::parse_from_bytes(packed.as_slice()) {
-                        Ok(smart_permissions) => smart_permissions,
-                        Err(err) => {
-                            return Err(ApplyError::InternalError(format!(
-                                "Cannot deserialize smart permission list: {:?}",
-                                err,
-                            )));
-                        }
-                    };
-
-                for smart_permission in smart_permissions.get_smart_permissions() {
-                    if smart_permission.name == name {
-                        return Ok(Some(smart_permission.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn set_smart_permission(
-        &mut self,
-        org_id: &str,
-        name: &str,
-        new_smart_permission: SmartPermission,
-    ) -> Result<(), ApplyError> {
-        let address = compute_smart_permission_address(org_id, name);
-        let d = self.context.get_state_entry(&address)?;
-        let mut smart_permission_list = match d {
-            Some(packed) => match protobuf::parse_from_bytes(packed.as_slice()) {
-                Ok(smart_permissions) => smart_permissions,
-                Err(err) => {
-                    return Err(ApplyError::InternalError(format!(
-                        "Cannot deserialize smart permission list: {}",
-                        err,
-                    )));
-                }
-            },
-            None => SmartPermissionList::new(),
-        };
-        // remove old smart_permission if it exists and sort the smart_permission by name
-        let smart_permissions = smart_permission_list.get_smart_permissions().to_vec();
-        let mut index = None;
-        let mut count = 0;
-        for smart_permission in smart_permissions.clone() {
-            if smart_permission.name == name {
-                index = Some(count);
-                break;
-            }
-            count = count + 1;
-        }
-
-        match index {
-            Some(x) => {
-                smart_permission_list.smart_permissions.remove(x);
-            }
-            None => (),
-        };
-        smart_permission_list
-            .smart_permissions
-            .push(new_smart_permission);
-        smart_permission_list
-            .smart_permissions
-            .sort_by_key(|sp| sp.clone().name);
-        let serialized = match protobuf::Message::write_to_bytes(&smart_permission_list) {
-            Ok(serialized) => serialized,
-            Err(_) => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot serialize smart permission list",
-                )));
-            }
-        };
-        self.context
-            .set_state_entry(address, serialized)
-            .map_err(|err| ApplyError::InternalError(format!("{}", err)))?;
-        Ok(())
-    }
-
-    pub fn delete_smart_permission(&mut self, org_id: &str, name: &str) -> Result<(), ApplyError> {
-        let address = compute_smart_permission_address(org_id, name);
-        let d = self.context.delete_state_entry(&address.clone())?;
-        let deleted = match d {
-            Some(deleted) => deleted,
-            None => {
-                return Err(ApplyError::InternalError(String::from(
-                    "Cannot delete smart_permission",
-                )));
-            }
-        };
-        if deleted != address {
-            return Err(ApplyError::InternalError(String::from(
-                "Cannot delete smart_permission",
-            )));
-        };
-        Ok(())
-    }
-
-    pub fn get_organization(&mut self, id: &str) -> Result<Option<Organization>, ApplyError> {
-        let address = compute_org_address(id);
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let orgs: OrganizationList = match protobuf::parse_from_bytes(packed.as_slice()) {
-                    Ok(orgs) => orgs,
-                    Err(err) => {
-                        return Err(ApplyError::InternalError(format!(
-                            "Cannot deserialize organization list: {:?}",
-                            err,
-                        )));
-                    }
-                };
-
-                for org in orgs.get_organizations() {
-                    if org.org_id == id {
-                        return Ok(Some(org.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-
-    pub fn get_agent(&mut self, public_key: &str) -> Result<Option<Agent>, ApplyError> {
-        let address = compute_agent_address(public_key);
-        let d = self.context.get_state_entry(&address)?;
-        match d {
-            Some(packed) => {
-                let agents: AgentList = match protobuf::parse_from_bytes(packed.as_slice()) {
-                    Ok(agents) => agents,
-                    Err(err) => {
-                        return Err(ApplyError::InternalError(format!(
-                            "Cannot deserialize record container: {:?}",
-                            err,
-                        )));
-                    }
-                };
-
-                for agent in agents.get_agents() {
-                    if agent.public_key == public_key {
-                        return Ok(Some(agent.clone()));
-                    }
-                }
-                Ok(None)
-            }
-            None => Ok(None),
-        }
-    }
-}
 
 pub struct SabreTransactionHandler {
     family_name: String,
@@ -989,7 +113,7 @@ impl TransactionHandler for SabreTransactionHandler {
             request.get_header().get_outputs()
         );
 
-        match payload.action {
+        match payload.get_action() {
             Action::CreateContract(create_contract_payload) => {
                 create_contract(create_contract_payload, signer, &mut state)
             }
